add dialog for delete button

+++ client/src/app/post/post-list/post-list.component.ts
@@ -1,5 +1,7 @@
 import { Component, OnInit } from '@angular/core';
+import { MatDialog, MatDialogRef } from '@angular/material';
 import { Router } from '@angular/router';
+import { DeletePostComponent } from '../delete-post/delete-post.component';
 import { PostService } from '../post.service';
 
 @Component({
@@ -9,7 +11,8 @@ import { PostService } from '../post.service';
 })
 export class PostListComponent implements OnInit {
   public posts;
-  constructor(private postService: PostService, private router: Router) {}
+
+  constructor(private postService: PostService, private dialog: MatDialog, private router: Router) {}
 
   ngOnInit() {
     this.getPosts();
@@ -30,10 +33,15 @@ export class PostListComponent implements OnInit {
   }
 
   deletePost(id: number) {
-    console.log('post to be deleted:' + id);
-    this.postService.deletePost(id);
-    this.reloadPage();
-
-    // this.reloadPage();
+    const dialogRef = this.dialog.open(DeletePostComponent);
+    dialogRef.afterClosed().subscribe(
+      result => {
+        if (result) {
+          this.postService.deletePost(id);
+          console.log('post to be deleted:' + id);
+        }
+      },
+      err => console.log('error while deleting' + err)
+    );
   }
 }
